Change how to sanitize file names
vim-zettel is used to fail to properly lowercase Turkish "ı" at the end of
file names. If it was not at the end, it is used to work right. I put an underscore
at end of filename when used `%title` option. Also now the plugin changes colons to
underscores and deletes any commas in file names.